Create Linked List
Problem statement- In an array, to insert the data or delete the data we have to shift each element. Processing these operations takes more time. So linked list is proposed. To insert we just need to add a new node and store is next part in the previous node. To delete the node we just need to change the next part to null.
In the code, I have created different functions for different operations-inserting the node at the beginning, insertion at the end, insertion at the specific position, deletion at the beginning, deletion at the end, deletion at the specific position, and displaying the linked list.