Create Day 6
The code is for reversing the words present in a string. First, we will split the string and convert them into an array of strings. Then using the StringBuilder class and reverse() method the words will get reversed. At last, we will concatenate the reversed word together to form the string.